Added methods escapeString + setOptionsList to lib; added tests
