+++ src/dropdown/test/index.js
@@ -5,7 +5,7 @@ process.env.NODE_ENV = 'test';
 import test from 'ava';
 import props from './../props';
 
-import { isNil, isNilOrEmptyStr, randomId, propLowerCase } from './../libs';
+import { isNil, isNilOrEmptyStr, randomId, propLowerCase, escapeString, setOptionsList } from './../libs';
 
 const element = '<dropdown-clab>';
 
@@ -116,4 +116,23 @@ test(`${element} propLowerCase`, t => {
   t.is(propLowerCase(), '');
   t.is(propLowerCase(obj, 'test'), '');
   t.is(propLowerCase(obj, 'foo'), 'bar');
+});
+
+test(`${element} escapeString`, t => {
+  const str = '+00 | Africa/Abidjan';
+  t.is(escapeString(str), '\\+00 \\| Africa\\/Abidjan');
+});
+
+test(`${element} setOptionsList`, t => {
+  const options = [
+    { value: 0, label: '+00 | Africa/Abidjan' }, 
+    { value: 1, label: '+01 | Europe/Rome' },
+    { value: 2, label: '-04 | America/Antigua' }
+  ];
+  const searchValue = 'rom';
+  const labelField = 'label';
+  t.deepEqual(setOptionsList(options, searchValue, labelField), [{ 
+    value: 1, 
+    label: '+01 | Europe/Rome' 
+  }]);
 });
